feat(UI-1198): organizations table - refactor delete modal
