+++ src/components/organisms/settings/organization/deleteModal.tsx
@@ -3,40 +3,43 @@ import React from "react";
 import { useTranslation } from "react-i18next";
 
 import { ModalName } from "@enums/components";
-import { useModalStore } from "@src/store";
+import { DeleteOrganizationModalProps } from "@interfaces/components";
 
-import { Button } from "@components/atoms";
+import { useModalStore } from "@store";
+
+import { Button, Loader } from "@components/atoms";
 import { Modal } from "@components/molecules";
 
-export const DeleteOrganizationModal = () => {
-	const { t } = useTranslation("settings", { keyPrefix: "organization" });
+export const DeleteOrganizationModal = ({ onDelete, isDeleting }: DeleteOrganizationModalProps) => {
+	const { t } = useTranslation("modals", { keyPrefix: "deleteOrganization" });
 	const { closeModal } = useModalStore();
-	const organizationName = useModalStore((state) => state.data as string);
+	const organization = useModalStore((state) => state.data) as { id: string; name: string };
 
 	return (
-		<Modal hideCloseButton name={ModalName.deleteOrganization}>
+		<Modal name={ModalName.deleteOrganization}>
 			<div className="mx-6">
-				<h3 className="mb-5 text-xl font-bold">{t("modal.deleteOrganization", { name: organizationName })}?</h3>
-				<p>{t("modal.line1")}</p>
-				<p>{t("modal.line2")}</p>
+				<h3 className="mb-5 text-xl font-bold">{t("deleteOrganization", { name: organization?.name })}?</h3>
+				<p className="font-light">{t("line1")}</p>
+				<p className="mt-1 font-light">{t("line2")}</p>
 			</div>
 
-			<div className="mt-8 flex w-full justify-end gap-2">
+			<div className="mt-14 flex justify-end gap-1">
 				<Button
-					ariaLabel={t("modal.cancel")}
-					className="bg-gray-1100 px-4 py-3 font-semibold"
+					ariaLabel={t("cancelButton")}
+					className="w-auto px-4 py-3 font-semibold hover:text-white"
 					onClick={() => closeModal(ModalName.deleteOrganization)}
-					variant="filled"
 				>
-					{t("modal.buttons.cancel")}
+					{t("cancelButton")}
 				</Button>
 
 				<Button
-					ariaLabel={t("modal.delete")}
-					className="px-4 py-3 font-semibold hover:bg-gray-1100 hover:text-white"
-					variant="outline"
+					ariaLabel={t("deleteButton")}
+					className="w-auto px-4 py-3 font-semibold hover:bg-error"
+					disabled={isDeleting}
+					onClick={() => onDelete(organization?.name, organization?.id)}
 				>
-					{t("modal.buttons.delete")}
+					{isDeleting ? <Loader size="sm" /> : null}
+					{t("deleteButton")}
 				</Button>
 			</div>
 		</Modal>
