+++ src/components/organisms/settings/user/organizations/index.ts
@@ -1,2 +1,1 @@
 export { UserOrganizationsTable } from "@components/organisms/settings/user/organizations/table";
-export { DeleteOrganizationModal } from "@components/organisms/settings/user/organizations/deleteOrganizationModal";
